+++ nym-vpn-android/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Improved account signup and subscription flow (https://github.com/nymtech/nym-vpn-client/pull/6014)
+
 ## [2026.12.0] - TBD
 
 ### Added